pyproject.toml
[project]
name = "www"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "cryptography>=46.0.0",
    "fastapi>=0.135.1",
    "pyyaml>=6.0.2",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.3.0",
    "uvicorn>=0.42.0",
]
